v0.6 - python3

+++ web_app/home/routes.py
@@ -19,7 +19,7 @@ blueprint = Blueprint(
 def index():
     counters = {
     'routers': len(Routers.query.all()),
-    'links': len(Links.query.all())/2,
+    'links': int(len(Links.query.all())/2),
     'prefixes': len(Prefixes.query.all()),
     'users': len(User.query.all())
     }
